[ANDROSDK-1949] Move LegendSetService to network
